refactor: move ENDPOINTS to constants file
