+++ packages/rest/lib/RestManager.ts
@@ -1,4 +1,5 @@
 import fetch from "node-fetch";
+import { ENDPOINTS } from "./constants";
 import {
     RestOptions,
     RequestMethods,
@@ -211,39 +212,3 @@ export class RestManager {
         return this.put(this.getFinalURL(ENDPOINTS.memberRole(userId, roleId)));
     }
 }
-
-export const ENDPOINTS = {
-    // Channel Endpoints
-    channelMessages: (channelId: string) => `/channels/${channelId}/messages`,
-    channelMessage: (channelId: string, messageId: string) => `/channels/${channelId}/messages/${messageId}`,
-
-    // Member Endpoints
-    memberNickname: (userId: string) => `/members/${userId}/nickname`,
-    memberRoles: (userId: string) => `/members/${userId}/roles`,
-
-    // Forum Endpoints
-    createForumThread: (channelId: string) => `/channels/${channelId}/forum`,
-
-    // List Endpoints
-    createListItem: (channelId: string) => `/channels/${channelId}/list`,
-
-    // Docs Endpoints
-    channelDocs: (channelId: string) => `/channels/${channelId}/docs`,
-    channelDoc: (channelId: string, docId: number) => `/channels/${channelId}/docs/${docId}`,
-
-    // Reactions Endpoints
-    channelReaction: (channelId: string, contentId: string, emoteId: number) => `/channels/${channelId}/content/${contentId}/emotes/${emoteId}`,
-
-    // Team XP Endpoints
-    memberXP: (userId: string) => `/members/${userId}/xp`,
-    roleXP: (userId: string) => `/roles/${userId}/xp`,
-
-    // Social Links Endpoints
-    getMemberSocialLinks: (userId: string, type: SocialLinkType) => `/members/${userId}/social-links/${type}`,
-
-    // Group Memberships Endpoints
-    groupMember: (groupId: string, userId: string) => `/groups/${groupId}/members/${userId}`,
-
-    // Role Memberships Endpoints
-    memberRole: (memberId: string, roleId: number) => `/members/${memberId}/roles/${roleId}`,
-};
