+++ packages/rest/lib/constants.ts
@@ -1,3 +1,5 @@
+import { SocialLinkType } from "./typings";
+
 /** The meanings of each of the known guilded api response status codes. */
 export const GUILDED_STATUS_CODES = {
     200: "The request was successful",
@@ -13,4 +15,40 @@ export const GUILDED_STATUS_CODES = {
     502: "Something went wrong on our end",
     503: "Something went wrong on our end",
     504: "Something went wrong on our end",
+};
+
+export const ENDPOINTS = {
+    // Channel Endpoints
+    channelMessages: (channelId: string) => `/channels/${channelId}/messages`,
+    channelMessage: (channelId: string, messageId: string) => `/channels/${channelId}/messages/${messageId}`,
+
+    // Member Endpoints
+    memberNickname: (userId: string) => `/members/${userId}/nickname`,
+    memberRoles: (userId: string) => `/members/${userId}/roles`,
+
+    // Forum Endpoints
+    createForumThread: (channelId: string) => `/channels/${channelId}/forum`,
+
+    // List Endpoints
+    createListItem: (channelId: string) => `/channels/${channelId}/list`,
+
+    // Docs Endpoints
+    channelDocs: (channelId: string) => `/channels/${channelId}/docs`,
+    channelDoc: (channelId: string, docId: number) => `/channels/${channelId}/docs/${docId}`,
+
+    // Reactions Endpoints
+    channelReaction: (channelId: string, contentId: string, emoteId: number) => `/channels/${channelId}/content/${contentId}/emotes/${emoteId}`,
+
+    // Team XP Endpoints
+    memberXP: (userId: string) => `/members/${userId}/xp`,
+    roleXP: (userId: string) => `/roles/${userId}/xp`,
+
+    // Social Links Endpoints
+    getMemberSocialLinks: (userId: string, type: SocialLinkType) => `/members/${userId}/social-links/${type}`,
+
+    // Group Memberships Endpoints
+    groupMember: (groupId: string, userId: string) => `/groups/${groupId}/members/${userId}`,
+
+    // Role Memberships Endpoints
+    memberRole: (memberId: string, roleId: number) => `/members/${memberId}/roles/${roleId}`,
 };
